Module stability_pool::contract

Functions

claim

Sends available claims to info.sender

deposit

Deposit debt tokens into the contract

distribute_funds

Calculate which and how much each user gets distributed from the liquidation

execute

get_distribution_ratios

Get user ratios of deposits from a list of deposits

instantiate

liquidate

Send repayments for the Positions contract. Positions contract sends back a distribute msg.

query

validate_assets

Note: This fails if an asset total is sent in two separate Asset objects. Both will be invalidated.

withdraw

Withdraw / Unstake, capital can be used for liquidations while unstaking